Fix handle_config()
Prior to this fix: When checking whether the parsed json value was null
or not, it would always act as if the value was not null. This is
because the pointer was not set as NULL in the first place. I've fixed
it by assigning it to NULL on declaration.